_TO_REVIEW/to_review_3/requirements.txt
#https://pypi.org/project/ofxparse/

ofxparse==0.21
pandas==2.2.3
openpyxl
langchain_groq==0.1.9
python-dotenv==1.0.1



#streamlit
#matplotlib
#seaborn
#plotly
#ipywidgets

# langchain_core==0.2.39

# gspread==6.1.4
# oauth2client==4.1.3


# google-api-python-client==1.7.9
# google-auth-httplib2==0.2.0
# google-auth-oauthlib==1.2.1

# #oauth2client==4.1.3
# google-auth==2.37.0
# google-auth-oauthlib==1.2.1
# google-auth-httplib2==0.2.0

# langchain_openai==0.1.23
# openai==1.45.0